+++ internal/lox/api/interpreter_api.go
@@ -1,24 +1,26 @@
-package lox
+package loxapi
 
 import (
 	"bytes"
+
+	"github.com/codecrafters-io/interpreter-tester/internal/lox"
 )
 
-func Run(source string) (string, int, string) {
-	ClearErrorFlags()
+func Evaluate(source string) (string, int, string) {
+	lox.ClearErrorFlags()
 	mockStdout := &bytes.Buffer{}
 	mockStderr := &bytes.Buffer{}
 
-	scanner := NewScanner(source)
+	scanner := lox.NewScanner(source)
 	tokens := scanner.ScanTokens(mockStdout, mockStderr)
-	parser := NewParser(tokens)
-	statements := parser.Parse(mockStdout, mockStderr)
-	Interpret(statements, mockStdout, mockStderr)
+	parser := lox.NewParser(tokens)
+	expression := parser.BasicParse(mockStdout, mockStderr)
+	lox.BasicInterpret(expression, mockStdout, mockStderr)
 
 	exitCode := 0
-	if HadParseError {
+	if lox.HadParseError {
 		exitCode = 65
-	} else if HadRuntimeError {
+	} else if lox.HadRuntimeError {
 		exitCode = 70
 	}
